Rename chart.js to base.js
